+++ HCClient/src/main/java/pages/HomePage.java
@@ -1,11 +1,10 @@
-package modules;
+package pages;
 
 import javafx.geometry.HPos;
 import javafx.geometry.Pos;
 import javafx.geometry.VPos;
 import javafx.scene.control.Label;
 import javafx.scene.layout.*;
-import javafx.scene.paint.Color;
 import userinterface.GUI;
 
 
@@ -27,18 +26,17 @@ public class HomePage {
         ColumnConstraints ccConst = new ColumnConstraints();
         ccConst.setPercentWidth(100);
         ccConst.setHalignment(HPos.CENTER);
-        homePage.setGridLinesVisible(true);
         homePage.getColumnConstraints().add(ccConst);
 
-        // get the usrname of thel ogged in user and construct the welcome label to scale
-        String user = GUI.loggedInUser.getUsername(); // retrieve this information from the database when the database is live
+        // get the username of thel ogged in user and construct the welcome label to scale
+        String user = guiInstance.loginInstance.loggedInUser.getUsername(); // retrieve this information from the database when the database is live
         Label welcomeLabel = GUI.scaleableText("Welcome to HuskyConnect, " + user, homePage.widthProperty(), homePage.heightProperty(), 50.);
         // place it in a box
         VBox labelBox = new VBox(welcomeLabel);
         labelBox.setAlignment(Pos.CENTER);
 
         // retrieve the user's feed
-        VBox userFeed = GUI.loggedInUser.generateUserFeed();
+        VBox userFeed = guiInstance.loginInstance.loggedInUser.generateUserFeed();
 
         // add both to the home page.
         // column 0, row 1 will set the welcome message to be 1/8 down the page
